pyproject.toml
[project]
name = "l0"
version = "0.1.0"
description = "Reliability layer for AI/LLM streaming with retry, guardrails, and observability"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
keywords = ["llm", "ai", "streaming", "reliability", "retry", "guardrails", "openai", "anthropic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "orjson>=3.9.0",
    "typing-extensions>=4.9.0",
    "uuid6>=2024.1.12",
]

[project.optional-dependencies]
openai = ["openai>=1.30"]
litellm = ["litellm>=1.40"]

observability = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-instrumentation-httpx>=0.45",
    "sentry-sdk>=1.40",
]

speed = [
    "uvloop>=0.19; platform_system!='Windows'",
]

dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "pytest-mock>=3.12",
    "ruff>=0.1",
    "mypy>=1.8",
    "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://github.com/your-org/l0-python"
Documentation = "https://github.com/your-org/l0-python#readme"
Repository = "https://github.com/your-org/l0-python"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/l0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::pydantic.warnings.PydanticDeprecatedSince20:litellm.*",
]

[tool.coverage.run]
source = ["l0"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
# Disable overly strict generic type parameter requirements
# Type inference handles these cases correctly
disallow_any_generics = false
mypy_path = "src"
exclude = ["src/"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_calls = false
warn_return_any = false
check_untyped_defs = false

[tool.basedpyright]
pythonVersion = "3.10"
typeCheckingMode = "strict"
reportMissingImports = "warning"
reportMissingTypeStubs = false
reportImportCycles = false
# Allow type inference to handle generic type arguments
reportMissingTypeArgument = false
# Suppress noise from optional dependencies
reportUnknownVariableType = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownParameterType = false
reportUnknownLambdaType = false
# Internal implementation details
reportPrivateUsage = false
reportUnnecessaryIsInstance = false
reportUnnecessaryCast = false
reportUnusedImport = "warning"
reportUnusedVariable = "warning"
reportUnusedFunction = "warning"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[dependency-groups]
dev = [
    "basedpyright>=1.35.0",
    "pytest-asyncio>=1.3.0",
]
